refactor: Rename example files and update references for clarity
- Renamed example files for the IoT Weather Station and Legal Contract to include numeric prefixes for better organization.
- Updated all relevant import paths and references in the README, main, tests, and web-main files to reflect the new naming convention.
- Ensured consistency across the examples directory for improved usability and clarity.
