+++ bindings/wasm/notarization_wasm/examples/README.md
@@ -77,10 +77,10 @@ The following examples are available:
 
 The following examples demonstrate practical use cases with proper field usage:
 
-| Name                                                                                                                                                       | Information                                                                        |
-| :--------------------------------------------------------------------------------------------------------------------------------------------------------- | :--------------------------------------------------------------------------------- |
-| [iot_weather_station](https://github.com/iotaledger/notarization/tree/main/bindings/wasm/notarization_wasm/examples/src/real-world/iot_weather_station.ts) | IoT weather station using dynamic notarization for continuous sensor data updates. |
-| [legal_contract](https://github.com/iotaledger/notarization/tree/main/bindings/wasm/notarization_wasm/examples/src/real-world/legal_contract.ts)           | Legal contract using locked notarization for immutable document hash attestation.  |
+| Name                                                                                                                                                          | Information                                                                        |
+| :------------------------------------------------------------------------------------------------------------------------------------------------------------ | :--------------------------------------------------------------------------------- |
+| [iot_weather_station](https://github.com/iotaledger/notarization/tree/main/bindings/wasm/notarization_wasm/examples/src/real-world/01_iot_weather_station.ts) | IoT weather station using dynamic notarization for continuous sensor data updates. |
+| [legal_contract](https://github.com/iotaledger/notarization/tree/main/bindings/wasm/notarization_wasm/examples/src/real-world/02_legal_contract.ts)           | Legal contract using locked notarization for immutable document hash attestation.  |
 
 <!--
