+++ bindings/wasm/notarization_wasm/examples/src/web-main.ts
@@ -9,8 +9,8 @@ import { updateState } from "./05_update_state";
 import { updateMetadata } from "./06_update_metadata";
 import { transferNotarization } from "./07_transfer_notarization";
 import { accessReadOnlyMethods } from "./08_access_read_only_methods";
-import { iotWeatherStation } from "./real-world/iot_weather_station";
-import { legalContract } from "./real-world/legal_contract";
+import { iotWeatherStation } from "./real-world/01_iot_weather_station";
+import { legalContract } from "./real-world/02_legal_contract";
 
 export async function main(example?: string) {
     // Extract example name.
